Commits 101
Activity: Your first commit
The following steps will guide you through the process of committing a change on GitHub.

Create a new file on this branch, in a _posts folder called 0000-01-02-yancheng0905.md. You can do so using this shortcut or manually as follows:
Return to the "Code" tab
In the branch drop-down, select "intro-slide-branch"
Click Create new file
In the "file name" field, type _posts/0000-01-02-yancheng0905.md. Entering the / in the filename will automatically place your file in the _posts directory.